+++ src/console/views/procs.rs
@@ -51,7 +51,9 @@ impl Pane for ProcsPane {
 
       let (status_char, status_color) = match task.status {
         TaskStatus::Running => ("●", color!("#4ec55e")),
-        TaskStatus::Down => ("○", color!("#e05a5a")),
+        TaskStatus::NotStarted => ("○", color!("#888888")),
+        TaskStatus::Exited(0) => ("○", color!("#5a8ee0")),
+        TaskStatus::Exited(_) => ("○", color!("#e05a5a")),
       };
       grid.draw_line(
         status_col,